+++ Fogger/src/main/java/dotblueshoes/fogger/GlobalFogEvent.java
@@ -9,7 +9,7 @@ import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.Entity;
 
-import dotblueshoes.fogger.ConfigHandler;
+import dotblueshoes.fogger.config.ConfigHandler;
 import dotblueshoes.fogger.Fogger;
 
 public class GlobalFogEvent {
@@ -24,32 +24,9 @@ public class GlobalFogEvent {
 
         /* Is the horizontal vertical. -1 stands for ceiling Fog. */
 		if (event.getFogMode() == 0) {
-			GlStateManager.setFogStart(visibleDistance * ConfigHandler.defaultFogDefinition.fogMinIntensity);
-			GlStateManager.setFogEnd(visibleDistance * ConfigHandler.defaultFogDefinition.fogMaxIntensity);
+			GlStateManager.setFogStart(visibleDistance * ConfigHandler.defaultFogMinIntensity);
+			GlStateManager.setFogEnd(visibleDistance * ConfigHandler.defaultFogMaxIntensity);
 		}
-
     }
-
-    // @SubscribeEvent
-	// public void colorFog(FogColors event) {
-
-    //     worldClient = Minecraft.getMinecraft().world;
-    //     entity = event.getEntity();
-    //     biomeName = worldClient.getBiome(new BlockPos(entity.posX, entity.posY, entity.posZ)).getRegistryName().toString();
-        
-    //     for (int i = 0; i < ConfigHandler.biomeFogs.length; i++)
-    //         if(biomeName.equals(ConfigHandler.biomeFogs[i].biomeName)) {
-    //             //event.setRed(ConfigHandler.biomeFogs[i].red);
-    //             //event.setGreen(ConfigHandler.biomeFogs[i].green);
-    //             //event.setBlue(ConfigHandler.biomeFogs[i].blue);
-    //             event.setRed(0F);
-    //             event.setGreen(0F);
-    //             event.setBlue(0F);
-    //          }
-
-    //     event.setGreen(0.5F);
-    //     event.setBlue(0.5F);
-    //     event.setRed(0.5F);
-	// }
-
+    
 }